+++ R/reexports.R
@@ -1,14 +1,4 @@
 
-# ----------------------- groupedstats -----------------------------------
-
-#' @export
-#' @importFrom groupedstats specify_decimal_p
-groupedstats::specify_decimal_p
-
-#' @export
-#' @importFrom groupedstats signif_column
-groupedstats::signif_column
-
 # ----------------------- tidyverse -------------------------------------
 
 #' @export